dns-proxy: Disable IPv6 DNS redirection rules
IPv6 DNS redirection rules are not correct, making
queries on IPv6 networks not working. This patch
disables IPv6 DNS redirection rules, allowing IPv6
queries to not go through the proxy and directly
to the name servers.

DNS proxy sets /etc/resolv.conf to point to the
system proxy IPv4 address. This means for local
processes (on dual stack and IPv6-only networks),
there will not be any IPv6 DNS queries to begin
with. However, the query will still be possible
as the proxy can query with IPv6.

With this patch, on ARC and Crostini, because
they do not refer to /etc/resolv.conf, IPv6
queries might be issued on dual stack or
IPv6-only networks. These queries will not go
through the proxy, but instead go directly to
the name servers.

BUG=b:201542631
TEST=dig google.com;
     sudo -u cups dig google.com;
     sudo -u chronos dig google.com;
     # Tried on dual stack and v6 only network.
TEST=Test DNS using ARC app;
TEST=Test DNS through Chrome;
TEST=tast run network.DNSProxy.*;
     tast run network.DNSProxyOverVPN.*;

Change-Id: I1f38763bed190f3f91cb21cca102c04fc4bb988c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3201665
Reviewed-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Jason Jeremy Iman <jasongustaman@chromium.org>
Tested-by: Jason Jeremy Iman <jasongustaman@chromium.org>