missive: Make missive/upload disconnectable.
DisconnectableClient is a wrapper on top of an abstract client which may
become available or unavailable at any moment. Calls made through the
client are rejected with error if the client is not available, and
cancelled immediately returning an error, if it becomes unavailable while
the call is in progress.
The wrapper is intended to be used for dBus connection to missive client
in Chrome and from daemon to Chrome in missived.
Switch to unavailable state at destruction, thus responding
CANCEL to all pending calls.

Use DisconnectableClient, so that dBus calls are rejected with error if
dBus is not available, and immediately cancel and return error by an
ongoing call if dBus was available but becomes unavailable now.

BUG=b:205814822
TEST=FEATURES="test" emerge-$BOARD missive

Change-Id: I9a0b89baba048a611a92d8bfe94eb3e62bdc8b5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3442547
Tested-by: Leonid Baraz <lbaraz@chromium.org>
Reviewed-by: Vignesh Shenvi <vshenvi@google.com>
Reviewed-by: Zach Trudo <zatrudo@google.com>
Commit-Queue: Zach Trudo <zatrudo@google.com>
(cherry picked from commit 39974fd6dab6c3769c2384ed9e43f72c961ab092)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3445445
Commit-Queue: Leonid Baraz <lbaraz@chromium.org>