shill: reset backoff when always-on VPN service is null.
This CL ensures the backoff mechanism implemented for always-on VPN is
reset when the always-on VPN service is set to null. It avoids the
ConnectAlwaysOnVpn() callback to crash because of a null service.

BUG=b:218005248
TEST=unit tests;
Set a VPN service as always-on VPN, let it connect and remove it. Repeat
multiple times.

Change-Id: I61081a14b09a5b893e94cc0f61637a639283cbcc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3644647
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Tested-by: Damien Dejean <damiendejean@google.com>
Commit-Queue: Damien Dejean <damiendejean@google.com>
(cherry picked from commit 9a0851f4917ab0a22a06d3a12399731c8c358d00)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3683238