cryptohome: Don't trigger TPM_REBOOT on wrong cred [M-103]
Fix TpmNotBoundToPcrAuthBlock to not return the TPM_REBOOT error when a
wrong password is supplied.

This fixes the issue that entering a wrong password on TPM 1.2 device
with old user accounts (that still use TpmBoundToPcrAuthBlock) started
to trigger the "Please restart your device to fix this error" error
screen. The same issue applied to the password change scenarios, because
after detecting the password change we enforce online login, but it
never completes because of this spurious "Please restart" screen.

BUG=b:238114321
TEST=create user on TPM 1.2 device with cryptohome hacked to never use TpmBoundToPcrAuthBlock, enter wrong password, verify no "Please restart your device to fix this error" appears

Change-Id: Ib733255097cfdccecdd12337c74ddf11abe1d033
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3751148
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: John L Chen <zuan@chromium.org>
Reviewed-by: Yi Chou <yich@google.com>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>