cros-disks: Try to unmount with default options first
When unmounting a mount point, try to unmount it in a gentle way with
default options first. If this fails with the EBUSY error, then try to
forcefully unmount it with the MNT_FORCE and MNT_DETACH options.

There is a suspicion that unmounting with the default options might
prevent some filesystem corruptions. This is not proven yet, but as a
precaution, we take this extra step before forcefully unmounting.

This CL reverts the effects of CL 3631745.

BUG=chromium:1360642
TEST=tast run dut platform.CrosDisks*

Change-Id: Iad9cfa465aa97986732bb48ade7287cdd3c03e7d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3866922
Reviewed-by: Marcello Salomao <msalomao@google.com>
Commit-Queue: François Degros <fdegros@chromium.org>
Tested-by: François Degros <fdegros@chromium.org>