cros-disks: Kill archive FUSE mounter when unmounting
This is a workaround for one of the symptoms of b/258344222. To prevent
the umount() syscall from blocking for too long, cros-disks requests the
termination of the FUSE process before calling umount(). This is only
effective if there is any such process, and if this process is marked as
"safe" to killed. In practice, it should only affect archive mounters.

BUG=b:258344222
TEST=tast run dut platform.CrosDisksArchive

Change-Id: Ica609985fe8263a9ba11b8d71a9008d233f42dbb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4035872
Tested-by: François Degros <fdegros@chromium.org>
Reviewed-by: Alexander Bolodurin <alexbn@chromium.org>
Commit-Queue: François Degros <fdegros@chromium.org>
(cherry picked from commit 539584dc37a6a907ef683f00bec12d3fdad26ff5)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4041045