init: Fix TpmOwned path
There was a spelling mistake in the TPMOwnedFile path. This path
is read to determine whether the TPM is owned when checking whether
the device needs to be clobbered even though there's no devmode file
present and boot is in verified mode. With the incorrect path, the file
can not be read and the default behavior is to assume that the tpm is owned.

BUG=b:172210918, b:263516339
TEST=emerge-hatch chromeos-init, cros deploy
before the change, in /var/log/messages there was an error:
Could not determine TPM owned, failed to read /sys/class/tpm/tmp0/device/owned.
No longer see this error with this fix.

Change-Id: I1b5f231aef93295456bd85063f921326b33ef9b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4121312
Tested-by: Sarthak Kukreti <sarthakkukreti@google.com>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@google.com>
Commit-Queue: Alexis Savery <asavery@chromium.org>
(cherry picked from commit ac53c223b182819813ca5195fe8f9729e6e9466f)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4134305
Commit-Queue: Sarthak Kukreti <sarthakkukreti@google.com>