modemfwd: Poll for modem health as a signal for modem recovery
This CL adds a mechanism to periodically poll the modem's primary port
to determine if the modem is healthy, and if the modem is in a bad state
for enough time (currently 1 minute), then the helpers' reboot call will
be exercised (currently only for fm350).

For now, the recovery is targeting the fm350 modem, but the health poll
is in place for all modems to start emitting relevant metrics to
identify stability issues and determine if recovery is necessary outside
the FM350.

There are several necessary considerations that have gone into the
design:

- The timer usage is seamless around suspend. Even though the modem goes
  into low power mode during suspend, it still responds to mbim
  commands, so the health check will pass if the modem is healthy around
  suspend.
- The timer is paused and the failure count is reset around all firmware
  flashes to prevent an opportunity for a recovery to be invoked during a
  firmware flash.
- Even if the modem were to disappear immediately before a suspend
  attempt, it takes 120 seconds for suspend to fail 10x and trigger
  shutdown. With a recovery trigger of 3x failures 20s apart, and a 25s
  recovery time, we should recovery in even the worst case. The
  alternative is for modemfwd to block suspend, which is more a more
  complex solution.

The above scenarios were tested, and the overall mechanism was tested by
manipulating the mbim-proxy to emulate an unresponsive modem via
`killall -STOP mbim-proxy` until recovery is triggered, and then
`killall -CONT mbim-proxy` to allow the cellular stack to re-initialize
after recovery. Detection and recovery have been tested thoroughly on
FM350, and detection has been tested on hatch (l850gl).

BUG=b:241850118
TEST=Tested the above line-items

Change-Id: I8760c9adcdf30576a30e51d8c0eb03bc2fd8de01
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4242426
Reviewed-by: Aashay Shringarpure <aashay@google.com>
Tested-by: Daniel Winkler <danielwinkler@google.com>
Commit-Queue: Daniel Winkler <danielwinkler@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
(cherry picked from commit baae28ff199dcc679c050858a535f2c2455ac296)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4395375
Reviewed-by: Madhav <madhavadas@google.com>