login: Fix liveness checker timeout.
What the original code did was:
- waitpid(2) for all chrome processes as many as possible.
- If the browser process is kept alive, send SIGABRT.
- waitpid(2) for all chrome processes again.
- If a process in the group still kept alive, send SIGKILL to all the
  processes.
- waitpid(2) for all chrome processes again.

On the other hand, we'd like to run HandleExit as a callback for the
process termination, because it handles process restarting. However,
depending on the timing, the browser process may be already waitpid(2)ed
in AbortAndKillAll(), so it may not be called.

This is the fix for the case. Now, we just send SIGABRT/SIGKILL,
but not waitpid, and wait by using OneShotTimer. Thus, if the process
is terminated as expected, HandlExit will be called.

BUG=b:268316763
TEST=Ran unittests. Also run following manually.
  touch /run/session_manager/enable_hang_detection && restart ui
  kill -STOP ${CHROME_PID}
  Then chrome should be restarted.

Change-Id: Ib7d61efc045bfb93d294673d1563f19969b23bda
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4289364
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
(cherry picked from commit 9b42443b19ed267546e19a6f0eaac79c2e024e6b)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4421259
Reviewed-by: Louis Rassaby <rassaby@google.com>
Commit-Queue: Xiyuan Xia <xiyuan@chromium.org>
Auto-Submit: Pablo Ceballos <pceballos@chromium.org>
Tested-by: Pablo Ceballos <pceballos@chromium.org>