dns-proxy: Don't remove file on empty nameservers
Calling "unlink()" causes crash due to seccomp. Instead of updating the
seccomp policy, fix the issue by not removing the file as it is not
necessary. The check is updated to only nameservers and the log is
updated to warning. This is done as it this can happen on (a) no network
or (b) IPv6 network with not yet ready nameservers.

This patch also removes very verbose writing to /etc/resolv.conf log.

BUG=b:275253650
TEST=Observe no crash with empty nameservers;
TEST=tast run <DUT> network.ResolvConf*

Change-Id: Ib30fbb507422d82b736ceb042dc5d07a2e156f21
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4371825
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Jason Iman <jasongustaman@chromium.org>
(cherry picked from commit 401d85e3f8edde61739884e463c8c1ac0b629f68)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4379702
Commit-Queue: Jason Iman <jasongustaman@chromium.org>