cryptohome: UssMigrator works with FileSystemKeyset
This CL modifies the migrator to receive FileSystemKeyset directly
instead of working on a VaultKeyset.

With the recent changes in UssMigration sessions's |vault_keyset_| is
not always available. MigrateToUssDuringUpdateVaultKeyset() attempting
to access |vault_keyset_| created crashes. To fix the issue this CL
changes the code to not rely on the existence of a decrypted VaultKeyset
before UpdateAuthFactor is called.

BUG=b:287012647
TEST=Unit tests pass.

Change-Id: I5d5ef2c78bbda872983dbfba00da0c8b7a52f1a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4631923
Reviewed-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: Betul Soysal <betuls@google.com>
Tested-by: Betul Soysal <betuls@google.com>
(cherry picked from commit 82458f2247bfb238a56829060408a8dbad1fcc50)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4638256