flex_id: update udevadm trigger used to wait for network devices
Update the udevadm trigger from `create` to `add`. Create was no longer
triggering the necessary behavior to wait for all network devices to be
loaded. This is a result of the udev uprev that was landed in m119.

BUG=b:304807538
TEST=cros deploy to affected DUT, inspect flex_id behavior

Change-Id: Ibb03f0ff98f5b7a3bab22f9076c3107d060cf253
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4935776
Commit-Queue: Justin Ouellette <justinoue@google.com>
Tested-by: Justin Ouellette <justinoue@google.com>
Reviewed-by: Ting-Hao Lin <tinghaolin@google.com>
Auto-Submit: Justin Ouellette <justinoue@google.com>